--nodbus, first draft for #1825
