+++ src/firejail/main.c
@@ -120,6 +120,7 @@ int arg_noprofile = 0; // use default.profile if none other found/specified
 int arg_memory_deny_write_execute = 0;		// block writable and executable memory
 int arg_notv = 0;	// --notv
 int arg_nodvd = 0; // --nodvd
+int arg_nodbus = 0; // -nodbus
 int login_shell = 0;
 
 
@@ -1111,7 +1112,7 @@ int main(int argc, char **argv) {
 		else if (strncmp(argv[i], "--protocol=", 11) == 0) {
 			if (checkcfg(CFG_SECCOMP)) {
 				if (cfg.protocol) {
-					fwarning("a protocol list is present, the new list \"%s\" will not be installed\n", argv[i] + 11);
+					fwarning("two protocol lists are present, \"%s\" will be installed\n", cfg.protocol);
 				}
 				else {
 					// store list
@@ -1734,6 +1735,8 @@ int main(int argc, char **argv) {
 			arg_notv = 1;
 		else if (strcmp(argv[i], "--nodvd") == 0)
 			arg_nodvd = 1;
+		else if (strcmp(argv[i], "--nodbus") == 0)
+			arg_nodbus = 1;
 
 		//*************************************
 		// network
