+++ src/firejail/sandbox.c
@@ -838,6 +838,13 @@ int sandbox(void* sandbox_arg) {
 	}
 
 	//****************************
+	// Session D-BUS
+	//****************************
+	if (arg_nodbus)
+		dbus_session_disable();
+
+
+	//****************************
 	// hosts and hostname
 	//****************************
 	if (cfg.hostname)
